# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlatPhorm Universal API",
    "version": "1.0.0",
    "description": "Base OpenApi specification for PlatPhorm Network including schemas, docs, mcp, and history tracing."
  },
  "servers": [
    {
      "url": "https://api.platphormnews.com",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/history": {
      "get": {
        "summary": "Retrieve session span history",
        "description": "Retrieves the distributed span history based on tracing headers.",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The session ID for the sequence of interactions."
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The trace ID for distributed span tracking."
          },
          {
            "in": "header",
            "name": "x-span-id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The current span ID in the trace."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of history spans."
          }
        }
      }
    },
...
  }
}